requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.5.2
    # via django
attrs==22.1.0
    # via jsonschema
backports-datetime-fromisoformat==1.0.0
    # via flattentool
btrees==4.10.1
    # via zodb
cached-property==1.5.2
    # via libcove
certifi==2022.9.24
    # via
    #   requests
    #   sentry-sdk
cffi==1.15.1
    # via persistent
charset-normalizer==2.1.1
    # via requests
contextlib2==21.6.0
    # via schema
dealer==2.1.0
    # via -r requirements.in
defusedxml==0.7.1
    # via odfpy
django==3.2.16
    # via
    #   -r requirements.in
    #   django-bootstrap3
    #   libcoveofds
    #   libcoveweb
django-bootstrap3==22.1
    # via libcoveweb
django-environ==0.9.0
    # via libcoveweb
et-xmlfile==1.1.0
    # via openpyxl
flattentool==0.17.2
    # via
    #   libcove
    #   libcoveweb
idna==3.4
    # via requests
ijson==3.1.4
    # via flattentool
jsonref==0.2
    # via
    #   flattentool
    #   libcove
jsonschema==3.2.0
    # via
    #   -r requirements.in
    #   libcove
libcove==0.27.0
    # via
    #   libcoveofds
    #   libcoveweb
libcoveofds==0.0.0
    # via -r requirements.in
libcoveweb==0.25.0
    # via
    #   -r requirements.in
    #   libcoveofds
lxml==4.9.1
    # via flattentool
odfpy==1.4.1
    # via flattentool
openpyxl==3.0.10
    # via flattentool
persistent==4.9.1
    # via
    #   btrees
    #   zodb
pycparser==2.21
    # via cffi
pyrsistent==0.18.1
    # via jsonschema
pytz==2022.4
    # via
    #   django
    #   flattentool
requests==2.28.1
    # via
    #   libcove
    #   libcoveweb
rfc3339-validator==0.1.4
    # via libcove
rfc3987==1.3.8
    # via libcove
schema==0.7.5
    # via flattentool
sentry-sdk==1.9.10
    # via -r requirements.in
six==1.16.0
    # via
    #   jsonschema
    #   rfc3339-validator
    #   zodb
sqlparse==0.4.3
    # via django
transaction==3.0.1
    # via zodb
uc-rfc6266-parser==0.1.0
    # via libcoveweb
urllib3==1.26.12
    # via
    #   requests
    #   sentry-sdk
werkzeug==1.0.1
    # via uc-rfc6266-parser
xmltodict==0.13.0
    # via flattentool
zc-lockfile==2.0
    # via zodb
zc-zlibstorage==1.2.0
    # via flattentool
zconfig==3.6.0
    # via zodb
zodb==5.7.0
    # via
    #   flattentool
    #   zc-zlibstorage
zodbpickle==2.4
    # via zodb
zope-interface==5.4.0
    # via
    #   btrees
    #   persistent
    #   transaction
    #   zc-zlibstorage
    #   zodb

# The following packages are considered to be unsafe in a requirements file:
# setuptools
